Feat/16 kafka (#19)

+++ src/main/java/com/example/emergencyassistb4b4/report/dto/ReportResponseDto.java
@@ -8,7 +8,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -32,9 +31,9 @@ public class ReportResponseDto {
 
     private final String gu;
 
-    private final BigDecimal locationLat; // 위도
+    private final Double locationLat; // 위도
 
-    private final BigDecimal locationLng; // 경도
+    private final Double locationLng; // 경도
 
     private final LocalDateTime updatedAt;
